backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# Machine learning models
statsmodels==0.14.0
pmdarima==2.0.4
xgboost==2.0.2
scikit-learn==1.3.2

# Background jobs
redis==5.0.1
rq==1.15.1

# Storage
boto3==1.29.7

# Database (for future use)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

